Revert "feat: add chakra ui (#4)"
This reverts commit 22c45dacf70635975f8dd7aca8a6fb58afb85f93.
