+++ package.json
@@ -8,17 +8,12 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "@chakra-ui/react": "^1.7.3",
-    "@emotion/react": "^11",
-    "@emotion/styled": "^11",
     "@next-auth/fauna-adapter": "^1.0.1",
     "faunadb": "^4.4.1",
-    "framer-motion": "^4",
     "next": "12.0.7",
     "next-auth": "^4.0.5",
     "react": "17.0.2",
-    "react-dom": "17.0.2",
-    "react-icons": "^4.3.1"
+    "react-dom": "17.0.2"
   },
   "devDependencies": {
     "@types/node": "16.11.12",
