+++ pages/_app.tsx
@@ -1,18 +1,10 @@
 import type { AppProps } from 'next/app'
 import { SessionProvider } from 'next-auth/react'
-import { ChakraProvider } from '@chakra-ui/react'
-
-import { theme } from '~/assets/theme'
-import { SidebarDrawerProvider } from '~/contexts/sidebar-drawer'
 
 function MyApp({ Component, pageProps }: AppProps) {
   return (
     <SessionProvider session={pageProps.session}>
-      <ChakraProvider theme={theme}>
-        <SidebarDrawerProvider>
-          <Component {...pageProps} />
-        </SidebarDrawerProvider>
-      </ChakraProvider>
+      <Component {...pageProps} />
     </SessionProvider>
   )
 }
