--- src/components/header/index.ts
@@ -1,2 +1,0 @@
-export { Header } from './header'
-export { Profile } from './profile'
