--- src/components/sidebar/index.ts
@@ -1,2 +1,0 @@
-export { ActiveLink } from './active-link'
-export { Sidebar } from './sidebar'
